+++ src/sub-clients/orders/model/OrderItemRequest.ts
@@ -1,4 +1,0 @@
-export interface OrderItemRequest {
-  id: string;
-  videoId: string;
-}