+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [26.0.0] - 2020-01-21
+
+- Add note field to the Cart and Order
+- The placeOrder function on OrdersClient takes PlaceOrderRequest object
+
 ## [25.0.1] - 2020-01-21
 
 - Add deliveryDate to Order